desacoplamento ListExercise
